tutorial/requirements.txt
decorator==4.0.11
Django==1.11.1
djangorestframework==3.6.2
djangorestframework-jwt==1.10.0
pygments=2.2.0
ebaysdk

